requirements.txt
google-cloud-bigquery>=3.25.0
google-cloud-storage>=2.16.0
google-cloud-aiplatform>=1.66.0
python-dotenv>=1.0.1
rapidfuzz>=3.9.7
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
google-genai>=0.6.0
